Changed to appropriate type for few functions.

+++ lib/heap.h
@@ -3,9 +3,9 @@
     #include <stdbool.h>
     #include "./vector.h"
     
-    void heapify(int, int *, bool);
+    void heapify(size_t, int *, bool);
     int heappush(Vector *, int, bool);
     int heappop(Vector *, bool);
     int heappushpop(Vector *, int, bool);
-    void decreasekey(int, int *, int, bool);
+    void decreasekey(size_t, int *, int, bool);
 #endif
